* logicalaxis_to_array:
  refact(MessageTypeToImage.cs): change type of output object to BitmapFrame refact(MessageTypeToImage.cs): read the images from the ResourceDictionary xaml file
  feat(IconsCollection.xaml): 新增Icon资源文件 refact(resources): 新增Root/Resources文件夹，存放图标、dll等外部资源 refact(MainWindow.xaml): Motion Control Panel的Show/Hide按钮自动生成 refact(LogicalMotionComponent.cs): DisplayName更名为Caption feat(LogicalMotionComponent.cs): 新增Icon属性，用来指定相关的窗口、按钮的显示图标 refact(ConfigLogicalMotionComponent.cs): 新增Icon属性 fix(IrixiEE0017.cs): Dispose()方法中增加_controller.CloseDevice() 方法用来关闭USB HID通讯
  feat(App.xaml.cs): 支持启动参数 refact(MainWindow.xaml, MainWindow.xaml.cs): 删除固定的DocumentPanel，在后台代码中根据LogicalMotionComponent来生成DocumentPanel feat(Splash window): 加入Splash Screen fix(SystemService.cs): Init() 函数中的for循环加入Task.Delay，解决程序启动时Main Window不能及时显示的问题 refact(SystemService.cs): Construction函数修改了LogicalAxis初始化的步骤，以支持动态LogicalAxis配置 refact(config): Configuration class 不再用于产生LogicaAxis和LogicalAligner，仅负责加载json配置项，传递到 LogicalMotionComponent class 和 LogicalAxis class 中 fix(IrixiEE0017.cs): Open() 函数中会检测ReadFWInfo() 的执行结果，如果读取固件信息失败，则初始化失败 refact(UserControls/Axis.xaml.cs): 删除 SystemService 属性，该属性包含在新的LogicalAxis Class里 feat(MotionComponentPanel): 新增该控件，用来根据json配置文件动态生成运动控制操作窗口（Type: DocumentPanel） refact(restore layout): workspace 布局在MainWindow的Loaded事件和Closing事件中恢复和保存 refact(restore layout): 新增--DefaultLayout启动参数，用于加载默认的workspace布局
  fix(SystemService.cs): fix MassMoveLogicalAxis() function with Task.WhenAny()
  feat(StepperControllerLib): add pca9534 read command
  refact(LuminosAxis.c): Init() 函数新增读取Luminos状态的功能

+++ Irixi_Aligner_Common/Interfaces/IAxis.cs
@@ -61,7 +61,7 @@ namespace Irixi_Aligner_Common.Interfaces
         /// <summary>
         /// Get the maximum stroke the axis supports
         /// </summary>
-        int MaxStroke { get; }
+        double MaxStroke { get; }
 
         /// <summary>
         /// Get or set the CW limitaion
